rn: reorg of apps dir; widgets dir; UI --> Layout;
