+++ apps/chess-rn/src/app/widgets/AppBarInDash.tsx
@@ -1,4 +1,4 @@
-import React, { PropsWithChildren } from 'react'
+import React from 'react'
 import { 
   View,
   StyleProp,
